+++ src/test/java/com/Delta/Pages/VacPackageDeals.java
@@ -1,4 +1,4 @@
-package Pages;
+package com.Delta.Pages;
 
 import com.Delta.Utilities.Driver;
 import org.openqa.selenium.WebElement;
